+++ dave/__init__.py
@@ -2,7 +2,7 @@
 # Kassel and individual contributors (see AUTHORS file for details). All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 
 from dave.archiv_io import *
 from dave.create import *