+++ dave/settings.py
@@ -8,7 +8,7 @@ import os
 # develop mode: running dave from ide and keycloak in local docker network
 # local mode: runing dave and keycloak in local docker network
 # production mode: running dave and keycloak on production server docker network
-stage = "local"  # set development modus
+stage = "production"  # set development modus
 if stage == "develop":
     # --- parameter for develop mode
     # keycloak settings